move privacy policy to SearchForm this was the only place it was being imported
