+++ src/components/RightPane/CoursePane/SearchForm/PrivacyPolicyBanner.js
